Fix: Refactor codes for ScriptFiltering.js and DictModal.js files To improve the structure and apply Styled-components appropriately Ref: #134
