+++ src/components/Header/Header.js
@@ -37,26 +37,15 @@ const Header = (props) => {
 
   const openModal = () => {
     setFadeOut(false);
-    // document.body.style.overflow = 'hidden';
     setLoginModal(true);
   };
 
   const closeModal = () => {
     setFadeOut(true);
-    // document.body.style.overflow = 'unset';
     setTimeout(() => {
       setLoginModal(false);
     }, 400);
   };
-
-  // useEffect(()=>{
-  //   if(modal_on){
-  //     document.body.style.overflow = 'hidden';
-  //   }
-  //   else if(!modal_on){
-  //     document.body.style.overflow = 'unset';
-  //   }
-  // },[modal_on])
 
   return (
     <>
