+++ src/components/NicknameModal/NicknameModal.js
@@ -79,13 +79,11 @@ const NicknameModal = (props) => {
 
   const openModal = () => {
     setFadeOut(false);
-    // document.body.style.overflow = "hidden";
     setNicknameModal(true);
   };
 
   const closeModal = () => {
     setFadeOut(true);
-    // document.body.style.overflow = "unset";
     setTimeout(() => {
       setNicknameModal(false);
     }, 400);
